new data plots using new sentiment analysis tool
We tried using a different sentiment analysis tool on the reviews to see if it would give us more usable data to plot that is more tailored to what we are doing. We used the SentiStrength tool (http://sentistrength.wlv.ac.uk) to get sentiment data and used the data in these plots. Because the SentiStrength tool outputs the positive and negative ratings of reviews as discrete values, some of these plots are boxplots.